+++ components/setup-link-instructions/auth0/3.mdx
