+++ lib/constants.ts
@@ -4,45 +4,56 @@ export const identityProviders = [
   {
     name: 'Auth0',
     id: 'auth0',
+    stepCount: 4,
   },
   {
     name: 'Azure',
     id: 'azure',
+    stepCount: 4,
   },
   {
     name: 'Generic SAML 2.0',
     id: 'generic-saml',
+    stepCount: 4,
   },
   {
     name: 'Google',
     id: 'google',
+    stepCount: 4,
   },
   {
     name: 'JumpCloud',
     id: 'jumpcloud',
+    stepCount: 4,
   },
   {
     name: 'Microsoft AD FS',
     id: 'microsoft-adfs',
+    stepCount: 4,
   },
   {
     name: 'Okta',
     id: 'okta',
+    stepCount: 4,
   },
   {
     name: 'OneLogin',
     id: 'onelogin',
+    stepCount: 4,
   },
   {
     name: 'OpenID Connect',
     id: 'openid-connect',
+    stepCount: 4,
   },
   {
     name: 'PingOne',
     id: 'pingone',
+    stepCount: 4,
   },
   {
     name: 'Rippling',
     id: 'rippling',
+    stepCount: 4,
   },
 ];
